Server by A. Slivko (fixed broadcast addr)

+++ src/libbcsmap/bcsmap.h
@@ -3,12 +3,10 @@
 // Note: this include is a beta feature for design- and compile-time
 #include "../liblinux_util/mscfix.h"
 
-#include <stddef.h>
 #include <stdint.h>
-//typedef unsigned char byte;
 
 typedef struct __map_t {
-	size_t width; // ширина карты - по горизонтали - количество столбцов
-	size_t height; // высота карты - по вертикали - количество строк
+	uint16_t width; // ширина карты - по горизонтали - количество столбцов
+	uint16_t height; // высота карты - по вертикали - количество строк
 	uint8_t *map_primitives;
 } BCSMAP;
